Change project structure and travis build
